pyproject.toml
[tool.poetry]
name = "taskreminder"
version = "0.1.0"
description = ""
authors = ["OLEkSandrKoniaiev <128646916+OLEkSandrKoniaiev@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.5.0"
django-dotenv = "^1.4.2"


[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
